Move local nav middleware to sub-app level rather than global level
The local nav middleware is only used within sub-apps so doesn't need
to be in the global level middleware folder.
