+++ src/apps/middleware.js
@@ -1,3 +1,5 @@
+const path = require('path')
+
 function setHomeBreadcrumb (name) {
   return function (req, res, next) {
     if (name) {
@@ -10,6 +12,20 @@ function setHomeBreadcrumb (name) {
   }
 }
 
+function setLocalNav (items = []) {
+  return function buildLocalNav (req, res, next) {
+    res.locals.localNavItems = items.map(item => {
+      const url = path.resolve(res.locals.CURRENT_PATH, item.path)
+      return Object.assign(item, {
+        url,
+        isActive: res.locals.CURRENT_PATH === url,
+      })
+    })
+    next()
+  }
+}
+
 module.exports = {
   setHomeBreadcrumb,
+  setLocalNav,
 }
